requirements_serve.txt
# LectureLens backend dependencies
# Install with: pip install -r requirements_serve.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9      
openai-whisper>=20231117
chromadb>=0.5.0
requests>=2.31.0
pydantic>=2.0.0
